Reduce (a lot of) allocations in ActiveRecord#cache_key
In the case where a record’s `updated_at` is only being used for a cache key we don’t need to convert it into a date time instance only to turn around and convert it back to a string. The only thing that matters is that when the record is updated the key returns a new string  that is previously not been used. In that case we can use the raw un-casted value of `updated_at` when generating the cache key. This has a surprisingly large effect on an application (CodeTriage) that has 50 cached partials:

Before: Total allocated: 743842 bytes (6626 objects)
After: Total allocated: 702955 bytes (6063 objects)

(743842 - 702955) / 743842.0 # => 5.4% ⚡️⚡️⚡️⚡️⚡️

This code was introduced in https://github.com/rails/rails/commit/aa8749eb52d7919a438940c9218cad98d892f9ad with no specific comment as to why `utc.to_s(:usec)` is required.

cc/ @dhh